backend/requirements.txt
flask
flask-cors
google-api-python-client
google-auth
google-auth-httplib2
python-dotenv
apscheduler
gunicorn
